Separate AddCasesModal content from modal trigger and extract useAddCasesInputs hook
